requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
pydantic
passlib[bcrypt]
